refactor: split the LSMR test suite by fixture and concern
lsmr/tests.rs was 1140 lines, the largest hand-written file in the
workspace, with its operators in three clumps interleaved with the tests
they serve.

Fixtures move to lsmr/fixtures.rs, a sibling of the modules under test so
both subtrees reach them without a cross-seam import. The tests split into
solve, breakdown and ladder. The one white-box bidiagonalization test —
the only one that drives GolubKahan instead of mlsmr — moves to
lsmr/bidiag/tests.rs and is renamed for what it checks.

Tests-only; no production code touched. All 35 bodies are byte-identical
to their originals.

Closes #258
